A handful of tiny improvements for the native runtime
  * Rewrite `BasicUtilities::monodroid_strsplit`
    Instead of reallocating the vector on each new token, pre-calculate
    the number of tokens and allocate enough memory for the entire
    vector at once.
  * Rewrite `init_logging_categories`
    Eliminate memory allocation by not using `monodroid_strsplit`,
    instead do all the processing using on-stack strings and their
    tokenizer.  Also, replace the `CATEGORY` preprocessor macro with a
    template function.

+++ src/monodroid/jni/xxhash.hh
@@ -52,7 +52,7 @@ namespace xamarin::android
 		force_inline static constexpr uint32_t hash (const char *input, size_t len) noexcept
 		{
 			return finalize (
-				(len >= 16 ? h16bytes<Seed> (input, len) : Seed + PRIME5) + len,
+				(len >= 16 ? h16bytes<Seed> (input, len) : Seed + PRIME5) + static_cast<uint32_t>(len),
 				(input) + (len & ~0xFU),
 				len & 0xF
 			);
